refactor(cog): add the use of Cogs for the entire bot
This is much cleaner and much easier to use on the long term.
It definitely asks quite some work to be done at first.
It allows much more scalability to add other commands or event
listeners. Cogs make it really easy to add or delete commands without
creating files of hundreds of lines.